update dependency target location
schemamama_rusqlite had a broken dependency, so we started to depending on a
fork that had that fixed. Since that fix is now merged upstream, this patch
makes us point to the official repository instead.

We're still pointing to the git repository instead of genericly refering to the
crate of the new version, since that crate wasn't published yet.